+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/ui/preferences/editors/AbstractRealNumberEditor.java
@@ -7,7 +7,7 @@ package net.sourceforge.pmd.eclipse.ui.preferences.editors;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Spinner;
 
-import net.sourceforge.pmd.properties.NumericPropertyDescriptor;
+import net.sourceforge.pmd.properties.PropertyDescriptor;
 import net.sourceforge.pmd.properties.PropertySource;
 
 /**
@@ -25,11 +25,14 @@ public abstract class AbstractRealNumberEditor<T extends Number> extends Abstrac
 
 
     protected final Spinner newSpinnerFor(Composite parent, PropertySource source,
-                                    NumericPropertyDescriptor<T> numDesc) {
+                                    PropertyDescriptor<T> numDesc) {
 
         Spinner spinner = newSpinnerFor(parent, DIGITS);
-        int min = (int) (numDesc.lowerLimit().doubleValue() * SCALE);
-        int max = (int) (numDesc.upperLimit().doubleValue() * SCALE);
+        //TODO: currently it is not possible to determine the numeric constraints values
+        //int min = (int) (numDesc.lowerLimit().doubleValue() * SCALE);
+        //int max = (int) (numDesc.upperLimit().doubleValue() * SCALE);
+        int min = (int) (DEFAULT_MINIMUM * SCALE);
+        int max = (int) (DEFAULT_MAXIMUM * SCALE);
         spinner.setMinimum(min);
         spinner.setMaximum(max);