+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/ui/preferences/editors/IntegerEditorFactory.java
@@ -12,10 +12,10 @@ import org.eclipse.swt.widgets.Spinner;
 
 import net.sourceforge.pmd.eclipse.ui.preferences.br.SizeChangeListener;
 import net.sourceforge.pmd.eclipse.ui.preferences.br.ValueChangeListener;
-import net.sourceforge.pmd.properties.IntegerProperty;
-import net.sourceforge.pmd.properties.NumericPropertyDescriptor;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
+import net.sourceforge.pmd.properties.PropertyFactory;
 import net.sourceforge.pmd.properties.PropertySource;
+import net.sourceforge.pmd.properties.constraints.NumericConstraints;
 
 /**
  * @author Brian Remedios
@@ -30,14 +30,9 @@ public class IntegerEditorFactory extends AbstractNumericEditorFactory<Integer> 
 
     public PropertyDescriptor<Integer> createDescriptor(String name, String description, Control[] otherData) {
 
-        return new IntegerProperty(
-            name,
-            description,
-            minimumIn(otherData).intValue(),
-            maximumIn(otherData).intValue(),
-            defaultIn(otherData).intValue(),
-            0.0f
-        );
+        return PropertyFactory.intProperty(name).desc(description)
+            .require(NumericConstraints.inRange(minimumIn(otherData).intValue(), maximumIn(otherData).intValue()))
+            .defaultValue(defaultIn(otherData).intValue()).build();
     }
 
 
@@ -49,7 +44,7 @@ public class IntegerEditorFactory extends AbstractNumericEditorFactory<Integer> 
     public Control newEditorOn(Composite parent, final PropertyDescriptor<Integer> desc, final PropertySource source,
                                final ValueChangeListener listener, SizeChangeListener sizeListener) {
 
-        final Spinner spinner = newSpinner(parent, (NumericPropertyDescriptor<Integer>) desc, valueFor(source, desc));
+        final Spinner spinner = newSpinner(parent, desc, valueFor(source, desc));
 
         spinner.addModifyListener(new ModifyListener() {
             public void modifyText(ModifyEvent event) {
@@ -68,12 +63,15 @@ public class IntegerEditorFactory extends AbstractNumericEditorFactory<Integer> 
     }
 
 
-    public static Spinner newSpinner(Composite parent, NumericPropertyDescriptor<?> desc, Object valueIn) {
+    public static Spinner newSpinner(Composite parent, PropertyDescriptor<?> desc, Object valueIn) {
 
         Spinner spinner = newSpinnerFor(parent, 0);
 
-        spinner.setMinimum(desc.lowerLimit().intValue());
-        spinner.setMaximum(desc.upperLimit().intValue());
+        //TODO: currently it is not possible to determine the numeric constraints values
+        //spinner.setMinimum(desc.lowerLimit().intValue());
+        //spinner.setMaximum(desc.upperLimit().intValue());
+        spinner.setMinimum(DEFAULT_MINIMUM);
+        spinner.setMaximum(DEFAULT_MAXIMUM);
 
         int value = valueIn == null ? spinner.getMinimum() : ((Number) valueIn).intValue();
         spinner.setSelection(value);