+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/ui/preferences/editors/MultiStringEditorFactory.java
@@ -7,6 +7,7 @@ package net.sourceforge.pmd.eclipse.ui.preferences.editors;
 import java.util.Collections;
 import java.util.List;
 
+import org.apache.commons.lang3.StringUtils;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Control;
@@ -16,10 +17,9 @@ import org.eclipse.swt.widgets.Text;
 
 import net.sourceforge.pmd.eclipse.ui.preferences.br.ValueChangeListener;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
+import net.sourceforge.pmd.properties.PropertyFactory;
 import net.sourceforge.pmd.properties.PropertySource;
-import net.sourceforge.pmd.properties.StringMultiProperty;
 import net.sourceforge.pmd.util.CollectionUtil;
-import net.sourceforge.pmd.util.StringUtil;
 
 /**
  * @author Brian Remedios
@@ -34,9 +34,7 @@ public class MultiStringEditorFactory extends AbstractMultiValueEditorFactory<St
 
     public PropertyDescriptor<List<String>> createDescriptor(String name, String optionalDescription,
                                                              Control[] otherData) {
-
-        return new StringMultiProperty(name, "String value "
-            + name, new String[] {""}, 0.0f, StringMultiProperty.DEFAULT_DELIMITER);
+        return PropertyFactory.stringListProperty(name).desc(optionalDescription).emptyDefaultValue().build();
     }
 
 
@@ -88,7 +86,7 @@ public class MultiStringEditorFactory extends AbstractMultiValueEditorFactory<St
     protected String addValueIn(Control widget, PropertyDescriptor<List<String>> desc, PropertySource source) {
 
         String newValue = ((Text) widget).getText().trim();
-        if (StringUtil.isEmpty(newValue)) {
+        if (StringUtils.isBlank(newValue)) {
             return null;
         }